refactor: city 패키지 이동
