+++ src/main/java/com/hubo/gillajabi/admin/application/dto/CityResponseDTO.java
@@ -1,7 +1,7 @@
 package com.hubo.gillajabi.admin.application.dto;
 
 import com.hubo.gillajabi.crawl.domain.constant.Province;
-import com.hubo.gillajabi.crawl.domain.entity.City;
+import com.hubo.gillajabi.city.domain.City;
 import lombok.AccessLevel;
 import lombok.AllArgsConstructor;
 import lombok.Getter;
