+++ src/main/java/com/hubo/gillajabi/city/domain/City.java
@@ -1,4 +1,4 @@
-package com.hubo.gillajabi.crawl.domain.entity;
+package com.hubo.gillajabi.city.domain;
 
 import com.hubo.gillajabi.crawl.domain.constant.Province;
 import com.hubo.gillajabi.crawl.infrastructure.dto.request.CityRequest;
