--- src/main/java/com/hubo/gillajabi/crawl/infrastructure/persistence/CityRepository.java
@@ -1,11 +1,0 @@
-package com.hubo.gillajabi.crawl.infrastructure.persistence;
-
-import com.hubo.gillajabi.crawl.domain.constant.Province;
-import com.hubo.gillajabi.crawl.domain.entity.City;
-import org.springframework.data.jpa.repository.JpaRepository;
-
-import java.util.Optional;
-
-public interface CityRepository extends JpaRepository<City, Long> {
-    Optional<City> findByNameAndProvince(String name, Province province);
-}
