+++ src/test/java/com/hubo/gillajabi/crawl/domain/entity/CityTest.java
@@ -1,5 +1,6 @@
 package com.hubo.gillajabi.crawl.domain.entity;
 
+import com.hubo.gillajabi.city.domain.City;
 import com.hubo.gillajabi.crawl.domain.constant.Province;
 import com.hubo.gillajabi.crawl.infrastructure.dto.request.CityRequest;
 import org.junit.jupiter.api.DisplayName;
